requirements.txt
openai
pandas
jinja2
pdfkit
python-dotenv
notebook
pdfplumber
chromadb
openai>=1.0.0
tiktoken
langchain
jinja2
typing
import hashlib
